# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title
  title: Api
  version: 1.0.0
  description: Cricket League Auction & Bidding Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/league-summary:
    get:
      operationId: getLeagueSummary
      tags: [analytics]
      summary: Get league-wide auction summary
      parameters:
        - name: auctionId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: League summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueSummary"
  /analytics/team-stats:
    get:
      operationId: getTeamStats
      tags: [analytics]
      summary: Get stats for all teams
      parameters:
        - name: auctionId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Team stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamStats"
  /analytics/bid-activity:
    get:
      operationId: getBidActivity
      tags: [analytics]
      summary: Get recent bid activity timeline
      parameters:
        - name: auctionId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Bid activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BidActivity"
  /analytics/player-pool:
    get:
      operationId: getPlayerPool
      tags: [analytics]
      summary: Get remaining player pool breakdown
      parameters:
        - name: auctionId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Player pool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerPool"
components:
  schemas:
...
    Player:
      type: object
      required: [id, name, country, category, basePrice, battingStyle, bowlingStyle, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        city:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        category:
          type: string
          enum: [batsman, bowler, all_rounder, wicket_keeper]
        battingStyle:
          type: string
        bowlingStyle:
          type: string
        basePrice:
          type: number
        status:
          type: string
          enum: [available, sold, unsold, in_auction]
        photoUrl:
          type: ["string", "null"]
        matches:
          type: ["integer", "null"]
        runs:
          type: ["integer", "null"]
        wickets:
          type: ["integer", "null"]
        strikeRate:
          type: ["number", "null"]
        economy:
          type: ["number", "null"]
        average:
          type: ["number", "null"]
        fifties:
          type: ["integer", "null"]
        hundreds:
          type: ["integer", "null"]
        rating:
          type: ["number", "null"]
        tags:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Team:
      type: object
      required: [id, name, shortName, primaryColor, purse, remainingPurse, maxPlayers, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        shortName:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        purse:
          type: number
        remainingPurse:
          type: number
        maxPlayers:
          type: integer
        maxOverseas:
          type: integer
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    LeagueSummary:
      type: object
      required: [totalPlayers, soldPlayers, unsoldPlayers, remainingPlayers, totalBids, highestBid, mostExpensivePlayer, highestSpendingTeam, avgSoldPrice, totalSpent]
      properties:
        totalPlayers:
          type: integer
        soldPlayers:
          type: integer
        unsoldPlayers:
          type: integer
        remainingPlayers:
          type: integer
        totalBids:
          type: integer
        highestBid:
          type: number
        mostExpensivePlayer:
          $ref: "#/components/schemas/Player"
        highestSpendingTeam:
          $ref: "#/components/schemas/Team"
        avgSoldPrice:
          type: number
        totalSpent:
          type: number
    TeamStats:
      type: object
      required: [team, playerCount, spent, remainingPurse, batsmen, bowlers, allRounders, wicketKeepers, overseas, budgetEfficiency]
      properties:
        team:
          $ref: "#/components/schemas/Team"
        playerCount:
          type: integer
        spent:
          type: number
        remainingPurse:
          type: number
        batsmen:
          type: integer
        bowlers:
          type: integer
        allRounders:
          type: integer
        wicketKeepers:
          type: integer
        overseas:
          type: integer
        budgetEfficiency:
          type: number
    BidActivity:
      type: object
      required: [id, playerName, teamName, teamColor, amount, createdAt]
      properties:
        id:
          type: integer
        playerName:
          type: string
        teamName:
          type: string
        teamColor:
          type: string
        amount:
          type: number
        createdAt:
          type: string
    PlayerPool:
      type: object
      required: [total, available, sold, unsold, byCategory]
      properties:
        total:
          type: integer
        available:
          type: integer
        sold:
          type: integer
        unsold:
          type: integer
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
    CategoryCount:
      type: object
      required: [category, total, available, sold]
      properties:
        category:
          type: string
        total:
          type: integer
        available:
          type: integer
        sold:
          type: integer