# lib/api-zod/src/generated/types/categoryCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cricket League Auction & Bidding Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface CategoryCount {
  category: string;
  total: number;
  available: number;
  sold: number;
}